refactor(sim): split 2000-line impl Simulation into topical sub-files
The `impl Simulation` block in `sim.rs` had grown to ~2000 lines
spanning 10 section-comment-delimited topics. `sim.rs` itself was
2385 lines — the single largest file in the crate.

This commit extracts each section into its own file under `sim/`,
following the pattern already established by `sim/construction.rs`,
`sim/lifecycle.rs`, and `sim/topology.rs`. The `Simulation` type is
unchanged; rustdoc still aggregates all 140 methods on one page.

New files:

* `sim/accessors.rs`     — world, tick, metrics, groups, stop lookup
* `sim/destinations.rs`  — imperative destination queue (push/clear/abort)
* `sim/eta.rs`           — ETA queries (single-stop + best-of-group)
* `sim/runtime.rs`       — set_max_speed / set_acceleration / set_deceleration
                            / set_weight_capacity / set_door_open_ticks etc.
* `sim/manual.rs`        — open_door / close_door / hold_door / set_target_velocity
                            / emergency_stop (ServiceMode::Manual commands)
* `sim/tagging.rs`       — per-entity metric tag assignment
* `sim/rider.rs`         — spawn_rider / build_rider / reroute /
                            settle_rider / despawn_rider + drain_events
* `sim/substep.rs`       — sub-tick stepping and phase hooks
* `sim/calls.rs`         — hall-call and car-call API

Remains in `sim.rs`:

* Struct definitions (`Simulation`, `ElevatorParams`, `LineParams`,
  `RiderBuilder`) and their inherent `impl` blocks that aren't
  `impl Simulation`.
* The `impl Simulation` "Internal helpers" section — private methods
  shared across the public-API sub-files.
* `impl fmt::Debug for Simulation`.
* The top-level helper `brake_target_stop`.

`sim.rs` drops from 2385 → 587 lines.

Four helpers needed visibility bumps from `fn` to `pub(super) fn`
since they are now called across the sub-module boundary:
`resolve_stop`, `require_elevator`, `emit_upgrade`,
`spawn_rider_inner`, `validate_positive_finite_f64`,
`validate_nonzero_u32`. No public API change.

All 604 lib tests + 156 doc tests pass, including
`api_surface_tests` which guards public-API regressions.
